[common/metatypes] refactor: move NodeInfo from common/management to common/metatypes
