+++ common/management/src/namespace/mod.rs
@@ -20,5 +20,4 @@ mod namespace_api;
 mod namespace_mgr;
 
 pub use namespace_api::NamespaceApi;
-pub use namespace_api::NodeInfo;
 pub use namespace_mgr::NamespaceMgr;
