+++ common/metatypes/Cargo.toml
@@ -7,9 +7,11 @@ publish = false
 edition = "2021"
 
 [dependencies]
+common-exception = {path = "../exception"}
 common-sled-store = {path = "../sled-store"}
 
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 async-raft = { git = "https://github.com/datafuse-extras/async-raft", tag = "v0.6.2-alpha.14" }
 
 [dev-dependencies]
