+++ query/src/api/http/v1/cluster_test.rs
@@ -21,7 +21,7 @@ use axum::AddExtensionLayer;
 use axum::Router;
 use common_base::tokio;
 use common_exception::Result;
-use common_management::NodeInfo;
+use common_metatypes::NodeInfo;
 use pretty_assertions::assert_eq;
 use tower::ServiceExt;
