+++ query/src/tests/context.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_management::NodeInfo;
+use common_metatypes::NodeInfo;
 
 use crate::clusters::Cluster;
 use crate::configs::Config;
